move active set generator to its own module
